Cleaning up plotting
